Create ListHeader component & refactor
Also expanded on changes in commit 4cfe3f1ab3df12ccb3a1d63c9deb6e1e6455ab48 so that the "SuperSetlist" is rendered when the user is not authenticated.
